+++ src/buffs/constants.ts
@@ -1,8 +1,9 @@
 export enum KNOWN_PROC_ID {
 	BurstHeal = '2',
-	Unknown = 'UNKNOWN_EFFECT_ID',
+	Unknown = 'UNKNOWN_PROC_EFFECT_ID',
 }
 
 export enum KNOWN_PASSIVE_ID {
 	TriggeredEffect = '66',
+	Unknown = 'UNKNOWN_PASSIVE_EFFECT_ID',
 }