+++ src/buffs/parsers/index.ts
@@ -1,3 +1,4 @@
 export { default as convertProcEffectToBuffs } from './convertProcEffectToBuffs';
+export { default as convertPassiveEffectToBuffs } from './convertPassiveEffectToBuffs';
 export { BuffSource, BuffStackType } from './buff-types';
 export type { IBuff, IUnitState, IEffectToBuffConversionContext } from './buff-types';